add Protecting Video Routes and fixed some bugs on images, still have bug playing videos

+++ controllers/videoController.js
@@ -23,7 +23,6 @@ export const search = async (req, res) => {
   } catch (error) {
     console.log(error);
   }
-  console.log(searchingBy);
   res.render("Search", { pageTitle: "Search", searchingBy, videos });
 };
 
@@ -53,7 +52,7 @@ export const videoDetail = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id).populate("creator"); // find vidoe from db, then find user detail only USING Object ID!
-    console.log(video);
+
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -66,9 +65,13 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (`${video.creator}` !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
-    console.log(error, "hi");
+    console.log(error);
     res.redirect(routes.home);
   }
 };
@@ -92,7 +95,12 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (`${video.creator}` !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
